To be rebased:
Other computer has unpushed refactor of event handling for keypresses, but
I haven't access right now, and need to work on this stuff. So in this branch
I'm doing *something* to pass keypress events to a fn along with updated zipper,
and future commits in this branch will be about actually using that data to
call the actions in the huet namespace.